Add obj IO compatibility
The new keywords still need to be tested, but everything does compile!
I'll have to either make a dummy scene, and check that the resulting
file data is correct, or just keep programming until I can render basic
triangles. The first would be preferrable, but we'll see if I have the
time.