Add? Transparency to all surfaces.
Some things of note:
High eta's look weird
1 looks awesome on a sphere, but doesn't refract
Cylinders look AWESOME
Reflections are intentionally faint, they are just a visual cue, not a
real thing to look at.
Triangles are still "messed up" when placed in the form of a box.
This commit also has numerous bug fixes that happened as I was trying to
fix the triangle thing.